Ultimos detalhes.

+++ src/model/Item.java
@@ -33,7 +33,7 @@ public class Item {
         return quantidade;
     }
 
-    public void setQuantidade(float quantidade) {
+    public void setQuantidade(double quantidade) {
         this.quantidade = quantidade;
     }
